+++ lib/src/config/gesture_config/gesture_config.g.dart
@@ -34,9 +34,11 @@ abstract class _$GestureConfigCWProxy {
 
   GestureConfig isManualSetZoomRect(bool isManualSetZoomRect);
 
+  GestureConfig panClaimRatio(double panClaimRatio);
+
   GestureConfig dragClaimSlopFactor(double dragClaimSlopFactor);
 
-  GestureConfig panClaimRatio(double panClaimRatio);
+  GestureConfig scaleClaimSlopFactor(double scaleClaimSlopFactor);
 
   /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored. You can also use `GestureConfig(...).copyWith.fieldName(...)` to override fields one at a time with nullification support.
   ///
@@ -58,8 +60,9 @@ abstract class _$GestureConfigCWProxy {
     int zoomStartMinDistance,
     int zoomSpeed,
     bool isManualSetZoomRect,
-    double dragClaimSlopFactor,
     double panClaimRatio,
+    double dragClaimSlopFactor,
+    double scaleClaimSlopFactor,
   });
 }
 
@@ -119,12 +122,16 @@ class _$GestureConfigCWProxyImpl implements _$GestureConfigCWProxy {
       this(isManualSetZoomRect: isManualSetZoomRect);
 
   @override
+  GestureConfig panClaimRatio(double panClaimRatio) =>
+      this(panClaimRatio: panClaimRatio);
+
+  @override
   GestureConfig dragClaimSlopFactor(double dragClaimSlopFactor) =>
       this(dragClaimSlopFactor: dragClaimSlopFactor);
 
   @override
-  GestureConfig panClaimRatio(double panClaimRatio) =>
-      this(panClaimRatio: panClaimRatio);
+  GestureConfig scaleClaimSlopFactor(double scaleClaimSlopFactor) =>
+      this(scaleClaimSlopFactor: scaleClaimSlopFactor);
 
   @override
 
@@ -148,8 +155,9 @@ class _$GestureConfigCWProxyImpl implements _$GestureConfigCWProxy {
     Object? zoomStartMinDistance = const $CopyWithPlaceholder(),
     Object? zoomSpeed = const $CopyWithPlaceholder(),
     Object? isManualSetZoomRect = const $CopyWithPlaceholder(),
-    Object? dragClaimSlopFactor = const $CopyWithPlaceholder(),
     Object? panClaimRatio = const $CopyWithPlaceholder(),
+    Object? dragClaimSlopFactor = const $CopyWithPlaceholder(),
+    Object? scaleClaimSlopFactor = const $CopyWithPlaceholder(),
   }) {
     return GestureConfig(
       enableLongPress: enableLongPress == const $CopyWithPlaceholder()
@@ -207,14 +215,18 @@ class _$GestureConfigCWProxyImpl implements _$GestureConfigCWProxy {
           ? _value.isManualSetZoomRect
           // ignore: cast_nullable_to_non_nullable
           : isManualSetZoomRect as bool,
-      dragClaimSlopFactor: dragClaimSlopFactor == const $CopyWithPlaceholder()
-          ? _value.dragClaimSlopFactor
-          // ignore: cast_nullable_to_non_nullable
-          : dragClaimSlopFactor as double,
       panClaimRatio: panClaimRatio == const $CopyWithPlaceholder()
           ? _value.panClaimRatio
           // ignore: cast_nullable_to_non_nullable
           : panClaimRatio as double,
+      dragClaimSlopFactor: dragClaimSlopFactor == const $CopyWithPlaceholder()
+          ? _value.dragClaimSlopFactor
+          // ignore: cast_nullable_to_non_nullable
+          : dragClaimSlopFactor as double,
+      scaleClaimSlopFactor: scaleClaimSlopFactor == const $CopyWithPlaceholder()
+          ? _value.scaleClaimSlopFactor
+          // ignore: cast_nullable_to_non_nullable
+          : scaleClaimSlopFactor as double,
     );
   }
 }
@@ -253,9 +265,11 @@ GestureConfig _$GestureConfigFromJson(Map<String, dynamic> json) =>
           (json['zoomStartMinDistance'] as num?)?.toInt() ?? 5,
       zoomSpeed: (json['zoomSpeed'] as num?)?.toInt() ?? 1,
       isManualSetZoomRect: json['isManualSetZoomRect'] as bool? ?? false,
+      panClaimRatio: (json['panClaimRatio'] as num?)?.toDouble() ?? 2,
       dragClaimSlopFactor:
           (json['dragClaimSlopFactor'] as num?)?.toDouble() ?? 0.5,
-      panClaimRatio: (json['panClaimRatio'] as num?)?.toDouble() ?? 2,
+      scaleClaimSlopFactor:
+          (json['scaleClaimSlopFactor'] as num?)?.toDouble() ?? 1,
     );
 
 Map<String, dynamic> _$GestureConfigToJson(GestureConfig instance) =>
@@ -275,6 +289,7 @@ Map<String, dynamic> _$GestureConfigToJson(GestureConfig instance) =>
       'zoomStartMinDistance': instance.zoomStartMinDistance,
       'zoomSpeed': instance.zoomSpeed,
       'isManualSetZoomRect': instance.isManualSetZoomRect,
-      'dragClaimSlopFactor': instance.dragClaimSlopFactor,
       'panClaimRatio': instance.panClaimRatio,
+      'dragClaimSlopFactor': instance.dragClaimSlopFactor,
+      'scaleClaimSlopFactor': instance.scaleClaimSlopFactor,
     };